# server/src/openapi.yaml
openapi: 3.0.3
info:
  title: Practice CRM API
  version: 1.0.0
  description: >
    A real REST API for the Practice CRM training app, built for API-automation
    practice (Postman, RestAssured, pytest, curl, etc.). Records created here
    show up in the CRM UI on next navigation/reload of the relevant page — this
    API is the CRM's single source of truth.


    **Reset**: `POST /api/reset` restores the deterministic seed state at any
    time (no auth required — see that endpoint's description). Seed login:
    `admin@crm.com` / `Pass@123` (also `rep@crm.com`, `viewer@crm.com`,
    `sam@crm.com`, `priya@crm.com` — all deactivated except `priya@crm.com`,
    same password). **Passwords are stored in plaintext by design** — this is
    a training tool with no real user data, and the seed credentials are
    published intentionally for testers to use.


    **Free-tier note**: this server may be deployed on a free host that sleeps
    after inactivity; a cold start re-seeds automatically, so state resets
    after idle periods. This is expected, not a bug.
servers:
  - url: /api
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    pageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
...
  responses:
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Request failed validation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Conflict:
      description: Request conflicts with the resource's current state (illegal transition, confirmation required, etc.).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
...
paths:
...
  /accounts/{id}:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    get: { tags: [Accounts], responses: { '200': { description: OK }, '404': { $ref: '#/components/responses/NotFound' } } }
    put:
      tags: [Accounts]
      requestBody: { required: true, content: { application/json: { schema: { type: object } } } }
      responses: { '200': { description: Updated }, '404': { $ref: '#/components/responses/NotFound' } }
    delete:
      tags: [Accounts]
      summary: Delete an account. `?cascade=true` deletes dependent Contacts/Deals; default (false) unlinks them.
      parameters: [{ name: cascade, in: query, schema: { type: boolean, default: false } }]
      responses: { '204': { description: Deleted }, '404': { $ref: '#/components/responses/NotFound' } }
...
  /deals/{id}:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    get: { tags: [Deals], responses: { '200': { description: OK }, '404': { $ref: '#/components/responses/NotFound' } } }
    put:
      tags: [Deals]
      summary: Full update. Stage transitions into Closed Won/Lost auto-stamp closeDate to now.
      requestBody: { required: true, content: { application/json: { schema: { type: object } } } }
      responses: { '200': { description: Updated }, '404': { $ref: '#/components/responses/NotFound' } }
    delete:
      tags: [Deals]
      summary: Delete a deal. Closed Won deals require `?confirm=DELETE`.
      parameters: [{ name: confirm, in: query, schema: { type: string } }]
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /quotes/{id}:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    get: { tags: [Quotes], responses: { '200': { description: OK }, '404': { $ref: '#/components/responses/NotFound' } } }
    put:
      tags: [Quotes]
      requestBody: { required: true, content: { application/json: { schema: { type: object } } } }
      responses: { '200': { description: Updated }, '404': { $ref: '#/components/responses/NotFound' } }
    delete: { tags: [Quotes], responses: { '204': { description: Deleted }, '404': { $ref: '#/components/responses/NotFound' } } }
  /quotes/{id}/transition:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    post:
      tags: [Quotes]
      summary: >
        Move a quote's status, validated against the allowed transition map.
        Accepting a quote with a linked, non-closed deal auto-closes that
        deal as Closed Won (closeDate stamped to now).
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [status], properties: { status: { type: string } } }
      responses:
        '200': { description: Transitioned. }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /customFieldDefs:
    get:
      tags: [CustomFieldDefs]
      parameters: [{ $ref: '#/components/parameters/page' }, { $ref: '#/components/parameters/pageSize' }]
      responses: { '200': { description: OK } }
    post:
      tags: [CustomFieldDefs]
      requestBody: { required: true, content: { application/json: { schema: { type: object } } } }
      responses: { '201': { description: Created }, '422': { $ref: '#/components/responses/ValidationError' } }
  /customFieldDefs/{id}:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    get: { tags: [CustomFieldDefs], responses: { '200': { description: OK }, '404': { $ref: '#/components/responses/NotFound' } } }
    put:
      tags: [CustomFieldDefs]
      requestBody: { required: true, content: { application/json: { schema: { type: object } } } }
      responses: { '200': { description: Updated }, '404': { $ref: '#/components/responses/NotFound' } }
    delete:
      tags: [CustomFieldDefs]
      summary: Delete a field. Unlinks its value from every record of its module and strips it from any Layout.
      responses: { '204': { description: Deleted }, '404': { $ref: '#/components/responses/NotFound' } }
  /layouts:
    get:
      tags: [Layouts]
      parameters: [{ $ref: '#/components/parameters/page' }, { $ref: '#/components/parameters/pageSize' }]
      responses: { '200': { description: OK } }
    post:
      tags: [Layouts]
      summary: Upsert a layout by (module, target). `fieldIds` must belong to the given module.
      requestBody: { required: true, content: { application/json: { schema: { type: object } } } }
      responses: { '200': { description: Updated (existing) }, '201': { description: Created }, '422': { $ref: '#/components/responses/ValidationError' } }
  /layouts/{id}:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    get: { tags: [Layouts], responses: { '200': { description: OK }, '404': { $ref: '#/components/responses/NotFound' } } }
    put:
      tags: [Layouts]
      requestBody: { required: true, content: { application/json: { schema: { type: object } } } }
      responses: { '200': { description: Updated }, '404': { $ref: '#/components/responses/NotFound' } }
...